+++ packages/graph-plugins-shared/src/plugins/internals/loose.ts
@@ -28,20 +28,19 @@ export type LoosePluginData = {
   optionalDependsOn: LooseGraphPlugin[];
 };
 
-/**
- * `CoreEventMap` is intersected into the return type here (not just the
- * `events` param) to keep the "merge yourself" contract honest at the type
- * level: every plugin must hand back a hub that still carries everything
- * core (and upstream plugins) emit, not just its own events. dropping this
- * intersection would let a plugin's loosely-typed return shape claim to
- * satisfy `LooseGraphPlugin` while actually losing `CoreEventMap` from the
- * threaded hub, which is exactly the failure mode `ExtractEventMap` relies
- * on this type to rule out.
- */
 export type LooseGraphPlugin = (options: LooseGraphPluginOptions) => {
   controls: LoosePluginData['controls'];
+  // [1]
   events: EventHub<CoreEventMap & LoosePluginData['events']>;
   actions: GraphActions<any>;
   getters: GraphGetters<any>;
   onAfterInit?: () => void;
 };
+
+// [1] `CoreEventMap` is intersected into the return type (not just the events
+// param) to keep the "merge yourself" contract honest at the type level: every
+// plugin must hand back a hub that still carries everything core (and upstream
+// plugins) emit, not just its own events. dropping this intersection would let
+// a plugin's loosely-typed return shape claim to satisfy `LooseGraphPlugin`
+// while actually losing `CoreEventMap` from the threaded hub, which is exactly
+// the failure mode `ExtractEventMap` relies on this type to rule out.
